requirements.txt
Pillow
colorama